% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_alleles.R
\name{available_alleles}
\alias{available_alleles}
\title{Show available alleles for HLA-I or HLA-II and corresponding methods.}
\usage{
available_alleles(
  pre_type = c("Binding", "Processing", "Immuno"),
  HLA_type = c("I", "II"),
  pre_method
)
}
\arguments{
\item{pre_type}{Character, what kind of method to use, could be "Binging", "Processing", or "Immuno"}

\item{HLA_type, }{Character, HLA type need to be predicted, can be I or II (Processing and Immuno can only be I).}

\item{pre_method}{Character, which method to used for predicting MHC-peptide binding. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}
}
\value{
A dataframe (tibble) contains one column of available alleles.
}
\description{
Show available alleles for HLA-I or HLA-II and corresponding methods.
}
\examples{
available_alleles("Binding","I","ann")
}
