% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_len.R
\name{available_len}
\alias{available_len}
\title{Show available predicted peptide length for sepcific method and allele for MHC-I}
\usage{
available_len(pre_method, pre_allele)
}
\arguments{
\item{pre_method}{Character, which method used to do prodiction}

\item{pre_allele}{Character, which MHC-I allele need to be predicted}
}
\value{
A numberic vector
}
\description{
Show available predicted peptide length for sepcific method and allele for MHC-I
}
\examples{
available_len("comblib_sidney2008","HLA-A*30:01")
}
