% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batchpep_binding.R
\name{batchpep_binding}
\alias{batchpep_binding}
\title{Predict binding affinity of peptides with MHC in batch mode.}
\usage{
batchpep_binding(pep_file, mhc_type, pep_length, allele, pre_method)
}
\arguments{
\item{pep_file}{Character, the path of peptide file needed to be processed, each row refers to a peptide sequence.}

\item{mhc_type}{MHC class, can be MHC-I or MHC-II.}

\item{pep_length}{A numeric or character vector, indicating the length for which to make predictions. For MHC-I, the length can be 8-15, for MHC-II, the length can be 11-30 or asis (take the length of input sequence as the peptide length)}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}}

\item{pre_method}{Character, indicating the prediction method. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
Predict binding affinity of peptides with MHC in batch mode.
}
\examples{
res <- batchpep_binding(pep_file=system.file("extdata", "random.pep", package = "MHCbinding"),
                                  mhc_type="MHC-I",pep_length=c(9,10),allele="HLA-A*01:01",pre_method="ann")
}
