% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pep.R
\name{extractSeq}
\alias{extractSeq}
\title{Extract sepcific sub-squence of peptide}
\usage{
extractSeq(seq, pos, len, indel = FALSE)
}
\arguments{
\item{seq}{Character, the peptide sequence.}

\item{pos}{Numeric, the position of mutated amino acid.}

\item{len}{Numeric, the epitope length needed to predicted.}

\item{indel}{Logical, whether the type of mutation is indel.}
}
\value{
Character, the exacted sub-peptide sequence
}
\description{
Given a peptide, position of mutated amino acid and needed epitope length, exacted the proper range of amino acids that used for silde window prediction of epitope.
}
\examples{
extractSeq("GHAHKVPRRLLKAAR",pos=2,len=8,indel=FALSE)

}
