% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mhcbinding.R
\name{general_mhcbinding}
\alias{general_mhcbinding}
\title{General MCHbinding}
\usage{
general_mhcbinding(mhc_type = c("MHC-I", "MHC-II"), ...)
}
\arguments{
\item{mhc_type}{MHC class}

\item{...}{Other para passed to mhcIbing or mhcIIbind}
}
\value{
Dataframe
}
\description{
General MCHbinding
}
\examples{
general_mhcbinding(mhc_type = "MHC-I",peptide = "SLYNTVATLY",
                  allele = "HLA-A*01:01",length = "8",
                   pre_method = "netmhcpan_el")
}
