% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/immuno_gnn.R
\name{immuno_gnn}
\alias{immuno_gnn}
\title{Predict immunogenicity score of peptide using Immuno-GNN model}
\usage{
immuno_gnn(
  input_type = "single",
  input_file,
  pep,
  hla,
  immuno_gnn_env,
  immuno_gnn_path,
  num_thread,
  temp_dir
)
}
\arguments{
\item{input_type, }{multiple mode for inputing file containing peptide and hla columns, or single mode for
one pair of peptide-hla}

\item{input_file, }{csv file path for multiple mode of input_type}

\item{pep, }{peptides for single mode of input_type}

\item{hla, }{single hla allele for single mode of input_type}

\item{immuno_gnn_env, }{the conda environment of Immuno-GNN}

\item{immuno_gnn_path, }{the dir path of Immuno_gnn.py, aaindex1_pca.csv , pMHCDataset and last_model.pth}

\item{num_thread, }{the thread should be used for prediction}

\item{temp_dir, }{the path of temp dir}
}
\value{
A dataframe containing pep, hla and predicted immunogenicity
}
\description{
Predict immunogenicity score of peptide using Immuno-GNN model
}
\examples{
immuno_gnn(pep = c("GHAHKVPRRLLKAA","SLYNTVATLY"),
                     hla = "HLA-A01:01", immuno_gnn_env="dl_gpu",
                     immuno_gnn_path="./data-raw/",num_thread=20,temp_dir=tempdir())
}
