% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf2binding.R
\name{maf2binding}
\alias{maf2binding}
\title{Predict peptide MHC binding based on MAF file.}
\usage{
maf2binding(
  annovar_path,
  maf_path,
  need_allsamples = FALSE,
  need_samples,
  hla_type,
  pep_length,
  allele,
  pre_method,
  method_type,
  client_path,
  tmp_dir,
  num_thread,
  mhcflurry_env = "mhcflurry-env",
  mhcnuggets_env = "mhcnuggets",
  netchop_path,
  Immuno_IEDB_path,
  Immuno_Deepimmuno_path,
  Deepimmuno_env,
  MixMHCpred_path,
  PRIME_path,
  seq2neo_env,
  seq2neo_path
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{maf_path}{Character, the path of MAF file needed to be converted.}

\item{need_allsamples}{Logical. Whether need all samples when multi-sample MAF file is supplied.}

\item{need_samples}{Character, the needed sample when multi-sample MAF file is supplied.}

\item{hla_type}{HLA class, can be HLA-I or HLA-II.}

\item{pep_length}{A numeric or character vector, indicating the length for which to make predictions. For MHC-I, the length can be 8-15, for MHC-II, the length can be 11-30 or asis (take the length of input sequence as the peptide length)}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}}

\item{pre_method}{Character, indicating the prediction method. Available methods for HLA-I or HLA-II can be obtained by \code{\link{available_methods}}}

\item{method_type, }{which type prediction method used, could be "Binding", "Processing" or "Immuno"}

\item{client_path}{The path of local IEDB tools, used when setting get_method as client}

\item{tmp_dir}{temp dir}

\item{num_thread}{specify the number of threads to be used in annotation}

\item{mhcflurry_env, }{the installed conda environment of mhcflurry, default is "mhcflurry-env"}

\item{mhcnuggets_env, }{the installed conda environment of mhcnuggets, default is "mhcnuggets"}

\item{netchop_path, }{the installed netchop path}

\item{Immuno_IEDB_path, }{the installed IEDB immunogenicity tool}

\item{Immuno_Deepimmuno_path, }{the deepimmuno-cnn.py script path}

\item{Deepimmuno_env, }{the conda envrionment of Deepimmuno}

\item{MixMHCpred_path, }{the intalled Mixmhcpred path}

\item{PRIME_path, }{PRIME tool path}

\item{seq2neo_env, }{the conda env of seq2neo}

\item{seq2neo_path, }{the of \code{immuno_Prediction} dir of Seq2Neo}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
This function use \code{\link{maf2seq}} to exacted mutated "new peptide" from MAF file, and predicted the binding affinity of these peptide with specific MHC allele using \code{\link{general_mhcbinding}}
}
\examples{
test <- maf2binding(annovar_path = "~/software/annovar/annovar/",need_allsamples=TRUE,
                   maf_path = system.file("extdata", "test.maf", package = "TCAP"),
                   hla_type = "I",pep_length = c(9,10),
                   allele = c("HLA-A*0203"),pre_method = "DeepImmuno",tmp_dir=tempdir(),
                   num_thread=1,method_type = "Immuno",
                   Deepimmuno_env = "DeepImmuno",
                   Immuno_Deepimmuno_path = ""~/software/DeepImmuno/"")
}
