% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf2binding.R
\name{maf2pep}
\alias{maf2pep}
\title{Convert MAF file to mutated peptides}
\usage{
maf2pep(
  annovar_path,
  maf_path,
  need_allsamples = FALSE,
  need_samples,
  num_thread
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{maf_path}{Character, the path of MAF file needed to be converted.}

\item{need_allsamples}{Logical. Whether need all samples when multi-sample MAF file is supplied.}

\item{need_samples}{Character, the needed sample when multi-sample MAF file is supplied.}

\item{num_thread}{specify the number of threads to be used in annotation}
}
\value{
A dataframe containg wildtype and the mutated protein sequence.
}
\description{
This function extract neo-peptides produced from the mutation recorded in the MAF file using the scripts provided by annovar.
}
\details{
The command run is like: annotate_variation.pl -out out_dir -build hg19 anaovar_input.txt humandb/ -hgvs -dbtype ensGene
coding_change.pl common_driver.exonic_variant_function humandb/hg19_ensGene.txt humandb/hg19_ensGeneMrna.fa -includesnp -onlyAltering --outfile coding.txt
}
\examples{
pep <- maf2pep(annovar_path = annovar_path,maf_path = system.file("extdata", "test.maf", package = "MHCbinding"),
               need_allsamples = TRUE,num_thread=1)
}
