% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf2binding.R
\name{maf2seq}
\alias{maf2seq}
\title{Convert MAF file to mutated peptides and extract proper range of amino acids}
\usage{
maf2seq(
  annovar_path,
  maf_path,
  need_allsamples = TRUE,
  need_samples,
  len,
  num_thread
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{maf_path}{Character, the path of MAF file needed to be converted.}

\item{need_allsamples}{Logical. Whether need all samples when multi-sample MAF file is supplied.}

\item{need_samples}{Character, the needed sample when multi-sample MAF file is supplied.}

\item{len}{Numeric, the epitope length needed to predicted.}

\item{num_thread}{specify the number of threads to be used in annotation}
}
\value{
A dataframe containg wildtype and the mutated protein sequence, plus exacted sequence from mutated peptides.
}
\description{
This function use \code{\link{maf2pep}} to convert MAF file to mutated peptides, and use \code{\link{extractSeq}} to extract proper range of amino acids.
}
\examples{
maf2seq(annovar_path = annovar_path,maf_path = system.file("extdata", "test.maf", package = "MHCbinding"),
        need_allsamples = TRUE,len = 9,num_thread=1)
}
