% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcIbinding_client.R
\name{mhcbinding_client}
\alias{mhcbinding_client}
\title{Predicting HLA peptide binding}
\usage{
mhcbinding_client(
  client_path,
  peptide,
  allele,
  length,
  pre_method,
  tmp_dir,
  hla_type
)
}
\arguments{
\item{peptide}{A character vector of input protein sequence.}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}.}

\item{length}{A numeric or character vector, indicating the length for which to make predictions. For MHC-I, the length can be 8-14}

\item{pre_method}{Character, indicating the prediction method. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}

\item{tmp_dir}{Character, the temp dir}

\item{hla_type}{HLA type, I or II}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
This is the wrapped function for IEDB commond tools.
}
