% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcbinding.R
\name{mhcIbinding}
\alias{mhcIbinding}
\title{Predicting MHC-I binding}
\usage{
mhcIbinding(
  peptide = c("GHAHKVPRRLLKAAR", "LKAADASADADGSGSGSGSG"),
  allele = c("HLA-A*01:01", "HLA-A*03:01"),
  length = c(8, 9),
  pre_method = c("ann", "comblib_sidney2008", "consensus", "netmhccons",
    "netmhcpan_ba", "netmhcpan_el", "netmhcstabpan", "pickpocket", "recommended", "smm",
    "smmpmbec")
)
}
\arguments{
\item{get_method}{}
}
\value{

}
\description{
Predicting MHC-I binding
}
\examples{
test <- mhcIbinding(peptide = "SLYNTVATLY",
                    allele = "HLA-A*01:01",length = "8",
                    pre_method = "netmhcpan_el",get_method = "api")
}
