% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcbinding.R
\name{mhcIIbinding}
\alias{mhcIIbinding}
\title{Predicting MHC-II binding based on IEDB API}
\usage{
mhcIIbinding(
  peptide = c("GHAHKVPRRLLKAAR"),
  allele = c("DRB1*15:01"),
  length = 15,
  pre_method = c("recommended", "consensus", "netmhciipan", "nn_align", "smm_align",
    "comblib", "tepitope")
)
}
\arguments{
\item{peptide}{A character vector of input protein sequence.}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}.}

\item{length}{A numeric or character vector, indicating the length for which to make predictions.}

\item{pre_method}{Character, indicating the prediction method. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
This is the wrapped function for IEDB API, the full document can refer to http://tools.iedb.org/main/tools-api/
}
\examples{
dt <- mhcIIbinding(pre_method = "nn_align")
}
