% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcIbinding_client.R
\name{mhcIIbinding_client}
\alias{mhcIIbinding_client}
\title{Predicting MHC-II binding based on IEDB local tools.}
\usage{
mhcIIbinding_client(
  client_path,
  peptide = c("GHAHKVPRRLLKAAR", "LKAADASADADGSGSGSGSG"),
  allele = c("DRB1*15:01", "DPB1*04:01"),
  length = c(14, 15),
  pre_method = c("comblib", "consensus3", "IEDB_recommended", "netmhciipan_el",
    "netmhciipan_ba", "nn_align", "smm_align", "sturniolo"),
  tmp_dir = tempdir()
)
}
\arguments{
\item{peptide}{A character vector of input protein sequence.}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}.}

\item{length}{A numeric or character vector, indicating the length for which to make predictions. For MHC-II, the length can be 11-30}

\item{pre_method}{Character, indicating the prediction method. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}

\item{tmp_dir}{Character, the temp dir}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
This is the wrapped function for IEDB commond tools.
}
\examples{
test <- mhcIIbinding_client(peptide = "GHAHKVPRRLLKAAR",
                           allele = "DRB1*15:01",length = "14",
                           pre_method = "netmhciipan_ba",client_path="~/software/mhc_ii/",tmp_dir=tempdir())
}
