% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcIbinding_client.R
\name{netchop_processing}
\alias{netchop_processing}
\title{Predict proteasomal processing using NetChop}
\usage{
netchop_processing(pep, temp_dir, netchop_path)
}
\arguments{
\item{pep, }{character vector of peptide sequence}

\item{temp_dir, }{temp dir}

\item{netchop_path, }{installed IEDB Netchop path}
}
\value{
Dataframe containing 3 column: "amino_acid", "prediction_score", "sequence_id"
}
\description{
Predict proteasomal processing using NetChop
}
\examples{
netchop_processing(c("GHAHKVPRRLLKAA","SLYNTVATLY"),"~/tmp/","~/software/netchop/")
}
