% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_peptides.R
\name{random_peptides}
\alias{random_peptides}
\title{Generate random peptide sequences}
\usage{
random_peptides(len, n)
}
\arguments{
\item{len, }{the length of peptide}

\item{n, }{the number needed peptides}
}
\value{
A dataframe with one column containing the generated random peptides.
}
\description{
This function samples amino acid to form random peptide sequences from 20 common amino acids.
}
\examples{
random_peptides(len=9,n=10)
}
