% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcIbinding_client.R
\name{seq2neo_help}
\alias{seq2neo_help}
\title{seq2neo help function to get ic50 and tap}
\usage{
seq2neo_help(pep_len, allele, peptide, tmp_dir, netchop_path, client_path)
}
\arguments{
\item{pep_len}{peptide length}

\item{allele}{hla alleles}

\item{peptide}{peptide sequences}

\item{tmp_dir}{tmp dir}

\item{netchop_path}{the path of neochop}

\item{client_path}{the path of IEDB tools}
}
\value{
dataframe containg TAP and IC50
}
\description{
seq2neo help function to get ic50 and tap
}
\examples{
seq2neo_help(pep_len=c(8,9),allele=c("HLA-A*01:01","HLA-A*01:02"),
                       tmp_dir="~/tmp/", netchop_path="~/software/netchop/",
                       peptide=c("SLYNTVATLY","GHAHKVPR"),
                       client_path = "~/software/mhc_i/src/")
}
