% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcIbinding_client.R
\name{split_pep}
\alias{split_pep}
\title{Sliding peptide sequence}
\usage{
split_pep(pep_seq, req_len)
}
\arguments{
\item{pep_seq, }{Character vector, peptide sequence}

\item{req_len, }{numberic, the length of resulting sub-sequence}
}
\value{
Character vector of sub-sequence
}
\description{
Sliding peptide sequence
}
\examples{
split_pep(c("GHAHKVPRRLLKAA", "SLYNTVATLY"),8)
}
