% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2binding.R
\name{txt2pep}
\alias{txt2pep}
\title{Convert TXT file to mutated peptides}
\usage{
txt2pep(
  annovar_path,
  txt_path,
  genome_version = c("hg19", "hg38", "mm10", "mm9"),
  tmp_dir,
  num_thread
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{txt_path}{Character, the path of TXT file needed to be processed.}

\item{genome_version}{Character, which genome build version of the TXT file, can be hg19 or hg38}

\item{tmp_dir}{Character, the temp dir}

\item{num_thread}{specify the number of threads to be used in annotation}
}
\value{
A dataframe containg wildtype and the mutated protein sequence.
}
\description{
This function extract neo-peptides produced from the mutation recorded in the TXT file using the scripts provided by annovar.The txt file is actually
a annovar input file, and the first five space- or tab- delimited columns represent chromosome, start position, end position, the reference nucleotides and the observed nucleotides
}
\details{
The command run is like: annotate_variation.pl -out out_dir -build hg19 anaovar_input.txt humandb/ -hgvs -dbtype ensGene
coding_change.pl common_driver.exonic_variant_function humandb/hg19_ensGene.txt humandb/hg19_ensGeneMrna.fa -includesnp -onlyAltering --outfile coding.txt
}
\examples{
pep <- txt2pep(annovar_path = annovar_path,txt_path = system.file("extdata", "test_avinput.txt", package = "MHCbinding"),
               genome_version = "hg19",tmp_dir=tempdir(),num_thread=1)
}
