% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt2binding.R
\name{txt2seq}
\alias{txt2seq}
\title{Convert TXT file to mutated peptides and extract proper range of amino acids}
\usage{
txt2seq(
  annovar_path,
  txt_path,
  genome_version = c("hg19", "hg38"),
  len,
  tmp_dir,
  num_thread
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{txt_path}{Character, the path of TXT file needed to be converted.}

\item{genome_version}{Character, which genome build version of the TXT file, can be hg19 or hg38}

\item{len}{Numeric, the epitope length needed to predicted.}

\item{tmp_dir}{Character, the temp dir}

\item{num_thread}{specify the number of threads to be used in annotation}
}
\value{
A dataframe containg wildtype and the mutated protein sequence, plus exacted sequence from mutated peptides.
}
\description{
This function use \code{\link{txt2pep}} to convert TXT file to mutated peptides, and use \code{\link{extractSeq}} to extract proper range of amino acids.
}
\examples{
txt2seq(annovar_path = annovar_path,txt_path = system.file("extdata", "test_avinput.txt", package = "MHCbinding"),
        genome_version = "hg19",len = 9,tmp_dir=tempdir(),num_thread=1)
}
