% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pep.R
\name{vcf2annova}
\alias{vcf2annova}
\title{Convert vcf file format into input file format of annovar.}
\usage{
vcf2annova(
  annovar_path,
  vcf_path,
  out_file,
  need_allsamples = TRUE,
  need_samples
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{vcf_path}{Character, the path of VCF file needed to be converted.}

\item{out_file}{Character, the path of converted file.}

\item{need_allsamples}{Logical. Whether need all samples when multi-sample VCF file is supplied.}

\item{need_samples}{Character, the needed sample when multi-sample VCF file is supplied.}
}
\description{
This function called convert2annovar.pl script provided by \href{https://annovar.openbioinformatics.org/en/latest/}{Annovar} to convert VCF to input file of annovar.
}
\details{
The command run is like :convert2annovar.pl -format vcf4 example/ex2.vcf -outfile ex2.avinput -allsample -withfreq -include
}
\examples{
temp_file <- tempfile()
file.create(temp_file)
vcf2annova(annovar_path = annovar_path,vcf_path = system.file("extdata", "test_grch38.vcf", package = "MHCbinding"),
           out_file = temp_file,need_allsamples = FALSE,
           need_samples = "TUMOR")

}
