% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pep.R
\name{vcf2pep}
\alias{vcf2pep}
\title{Convert VCF file to mutated peptides}
\usage{
vcf2pep(
  annovar_path,
  vcf_path,
  genome_version = c("hg19", "hg38"),
  need_allsamples = TRUE,
  need_samples,
  num_thread
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{vcf_path}{Character, the path of VCF file needed to be converted.}

\item{genome_version}{Character, which genome build version of the VCF file, can be hg19 or hg38}

\item{need_allsamples}{Logical. Whether need all samples when multi-sample VCF file is supplied.}

\item{need_samples}{Character, the needed sample when multi-sample VCF file is supplied.}

\item{num_thread}{specify the number of threads to be used in annotation}
}
\value{
A dataframe containg wildtype and the mutated protein sequence.
}
\description{
This function extract neo-peptides produced from the mutation recorded in the VCF file using the scripts provided by annovar.
}
\details{
The command run is like: annotate_variation.pl -out out_dir -build hg19 anaovar_input.txt humandb/ -hgvs -dbtype ensGene
coding_change.pl common_driver.exonic_variant_function humandb/hg19_ensGene.txt humandb/hg19_ensGeneMrna.fa -includesnp -onlyAltering --outfile coding.txt
}
\examples{
pep <- vcf2pep(annovar_path = annovar_path,vcf_path = system.file("extdata", "test_grch38.vcf", package = "MHCbinding"),
               genome_version = "hg38",need_allsamples = FALSE,
               need_samples = "TUMOR",num_thread=1)
}
