% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pep.R
\name{vcf2seq}
\alias{vcf2seq}
\title{Convert VCF file to mutated peptides and extract proper range of amino acids}
\usage{
vcf2seq(
  annovar_path,
  vcf_path,
  genome_version = c("hg19", "hg38"),
  need_allsamples = TRUE,
  need_samples,
  len,
  num_thread
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{vcf_path}{Character, the path of VCF file needed to be converted.}

\item{genome_version}{Character, which genome build version of the VCF file, can be hg19 or hg38}

\item{need_allsamples}{Logical. Whether need all samples when multi-sample VCF file is supplied.}

\item{need_samples}{Character, the needed sample when multi-sample VCF file is supplied.}

\item{len}{Numeric, the epitope length needed to predicted.}

\item{num_thread}{specify the number of threads to be used in annotation}
}
\value{
A dataframe containg wildtype and the mutated protein sequence, plus exacted sequence from mutated peptides.
}
\description{
This function use \code{\link{vcf2pep}} to convert VCF file to mutated peptides, and use \code{\link{extractSeq}} to extract proper range of amino acids.
}
\examples{
vcf2seq(annovar_path = annovar_path,vcf_path = system.file("extdata", "test_grch38.vcf", package = "TCAP"),
        genome_version = "hg38",
        need_allsamples = FALSE,need_samples = "TUMOR",len = 9,num_thread=1)
}
