% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_alleles.R
\name{available_alleles}
\alias{available_alleles}
\title{Show available alleles for MHC-I or MHC-II and corresponding methods.}
\usage{
available_alleles(pre_type = c("MHC-I", "MHC-II"), pre_method)
}
\arguments{
\item{pre_type}{Character, which MHC type need to be predicted, can be MHC-I or MHC-II}

\item{pre_method}{Character, which method to used for predicting MHC-peptide binding. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}
}
\value{
A dataframe (tibble) contains one column of available alleles.
}
\description{
Show available alleles for MHC-I or MHC-II and corresponding methods.
}
\examples{
available_alleles("MHC-I","ann")
}
