% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_methods.R
\name{available_methods}
\alias{available_methods}
\title{Show available prediction methods for MHC-I or MHC-II}
\usage{
available_methods(
  get_methods = c("api", "client"),
  pre_type = c("MHC-I", "MHC-II")
)
}
\arguments{
\item{get_methods}{Character, which method to used for predicting MHC-peptide binding, can be api, refering to IEDB API and client, refering to local client}

\item{pre_type}{Character, which MHC type need to be predicted, can be MHC-I or MHC-II}
}
\value{
A dataframe (tibble) contains two columns of method and corresponding version.
}
\description{
Show available prediction methods for MHC-I or MHC-II
}
\examples{
available_methods("api","MHC-I")
}
