% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_methods.R
\name{available_methods}
\alias{available_methods}
\title{Show available prediction methods for HLA-antigen binding, antigen processing or Immunogenicity.}
\usage{
available_methods(pre_type = c("Binding", "Processing", "Immuno"))
}
\arguments{
\item{pre_type}{Character, which kind of predicted method, could be Binding, Processing, or Immuno}
}
\value{
A dataframe (tibble) contains two columns of method and corresponding version.
}
\description{
Show available prediction methods for HLA-antigen binding, antigen processing or Immunogenicity.
}
\examples{
available_methods("Binding")
}
