% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available_methods.R
\name{available_methods}
\alias{available_methods}
\title{Show avaliable prediction methods for MHC-I or MHC-II}
\usage{
available_methods(
  get_methods = c("api", "client"),
  pre_type = c("MHC-I", "MHC-II")
)
}
\arguments{
\item{pre_type}{}
}
\value{

}
\description{
Show avaliable prediction methods for MHC-I or MHC-II
}
