% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_mhcbinding.R
\name{general_mhcbinding}
\alias{general_mhcbinding}
\title{General MCHbinding}
\usage{
general_mhcbinding(
  hla_type = c("I", "II"),
  client_path,
  peptide,
  allele,
  length,
  pre_method,
  tmp_dir,
  method_type,
  mhcflurry_type = "mt",
  mhcflurry_env = "mhcflurry-env",
  mhcnuggets_env = "mhcnuggets",
  netchop_path,
  Immuno_IEDB_path,
  Immuno_Deepimmuno_path,
  Deepimmuno_env
)
}
\arguments{
\item{hla_type}{HLA class}

\item{client_path}{The path of local IEDB tools, used when setting get_method as client}

\item{...}{Other para passed to mhcIbing or mhcIIbind}
}
\value{
Dataframe
}
\description{
General MCHbinding
}
\examples{
general_mhcbinding(hla_type = "I",peptide = "SLYNTVATLY",
                   allele = "HLA-A*01:01",length = "8",
                   pre_method = "netmhcpan_el",
                   client_path="~/software/mhc_i/src/",tmp_dir=tmp())
}
