% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcIbinding_client.R
\name{mhcbinding_client}
\alias{mhcbinding_client}
\title{Predicting HLA peptide binding}
\usage{
mhcbinding_client(
  client_path,
  peptide,
  allele = NULL,
  length,
  pre_method,
  tmp_dir,
  hla_type,
  method_type,
  mhcflurry_type = "mt",
  mhcflurry_env = "mhcflurry-env",
  mhcnuggets_env = "mhcnuggets",
  netchop_path,
  Immuno_IEDB_path
)
}
\arguments{
\item{peptide}{A character vector of input protein sequence.}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}.}

\item{length}{A numeric or character vector, indicating the length for which to make predictions. For MHC-I, the length can be 8-14}

\item{pre_method}{Character, indicating the prediction method. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}

\item{tmp_dir}{Character, the temp dir}

\item{hla_type}{HLA type, I or II}

\item{method_type, }{which type prediction method used, could be "Binding", "Processing" or "Immuno"}

\item{mhcflurry_type, }{is calculating wt peptide for mhcflurry}

\item{netchop_path, }{the path of Netchop}

\item{Immuno_IEDB_path, }{the path of IEDB Class I Immunogenicity tool}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
This is the wrapped function for IEDB commond tools.
}
