% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcbinding.R
\name{mhcIbinding}
\alias{mhcIbinding}
\title{Predicting MHC-I binding based on IEDB API}
\usage{
mhcIbinding(
  peptide = c("GHAHKVPRRLLKAAR", "LKAADASADADGSGSGSGSG"),
  allele = c("HLA-A*01:01", "HLA-A*03:01"),
  length = c(8, 9),
  pre_method = c("ann", "comblib_sidney2008", "consensus", "netmhccons",
    "netmhcpan_ba", "netmhcpan_el", "netmhcstabpan", "pickpocket", "recommended", "smm",
    "smmpmbec")
)
}
\arguments{
\item{peptide}{A character vector of input protein sequence.}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}.}

\item{length}{A numeric or character vector, indicating the length for which to make predictions. For MHC-I, the length can be 8-15}

\item{pre_method}{Character, indicating the prediction method. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
This is the wrapped function for IEDB API, the full document can refer to http://tools.iedb.org/main/tools-api/
}
\examples{
test <- mhcIbinding(peptide = "SLYNTVATLY",
                    allele = "HLA-A*01:01",length = "8",
                    pre_method = "netmhcpan_el")
}
