% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcIbinding_client.R
\name{mhcIbinding_client}
\alias{mhcIbinding_client}
\title{Predicting MHC-I binding based on IEDB local tools.}
\usage{
mhcIbinding_client(
  client_path,
  peptide = c("GHAHKVPRRLLKAA", "SLYNTVATLY"),
  allele = c("HLA-A*01:01", "HLA-A*03:01"),
  length,
  pre_method = c("ann", "comblib_sidney2008", "consensus", "netmhccons",
    "netmhcpan_ba", "netmhcpan_el", "netmhcstabpan", "pickpocket", "IEDB_recommended",
    "smm", "smmpmbec", "mhcflurry", "mhcnuggets", "IEDB", "PRIME2.0", "DeepImmuno",
    "Seq2Neo-CNN", "Netchop", "NetCTLpan"),
  tmp_dir = tempdir(),
  method_type = c("Binding", "Processing", "Immuno"),
  mhcflurry_type = "mt",
  mhcflurry_env = "mhcflurry-env",
  mhcnuggets_env = "mhcnuggets",
  netchop_path,
  Immuno_IEDB_path,
  Immuno_Deepimmuno_path,
  Deepimmuno_env
)
}
\arguments{
\item{peptide}{A character vector of input protein sequence.}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}. For \code{Netchop}, allele is not needed.}

\item{length}{A numeric or character vector, indicating the length for which to make predictions. For MHC-I, the length can be 8-14}

\item{pre_method}{Character, indicating the prediction method. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}

\item{tmp_dir}{Character, the temp dir}

\item{method_type, }{which type prediction method used, could be "Binding", "Processing" or "Immuno"}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
This is the wrapped function for IEDB commond tools.
}
\examples{
test <- mhcIbinding_client(peptide = "SLYNTVATLY",
                           allele = "HLA-A*01:01",length = "8",
                           pre_method = "netmhcpan_el",client_path="~/software/mhc_i/src/",tmp_dir=tempdir())
}
