% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhcbinding.R
\name{mhcIIbinding}
\alias{mhcIIbinding}
\title{Predicting MHC-II binding}
\usage{
mhcIIbinding(
  peptide = c("GHAHKVPRRLLKAAR", "LKAADASADADGSGSGSGSG"),
  allele = c("HLA-DRB1*01:01", "HLA-A*03:01"),
  length = c(8, 9),
  pre_method = c("ann", "comblib_sidney2008", "consensus", "netmhccons",
    "netmhcpan_ba", "netmhcpan_el", "netmhcstabpan", "pickpocket", "recommended", "smm",
    "smmpmbec")
)
}
\arguments{
\item{pre_method}{}
}
\value{

}
\description{
Predicting MHC-II binding
}
