% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2pep.R
\name{vcf2binding}
\alias{vcf2binding}
\title{Predict peptide MHC binding based on VCF file.}
\usage{
vcf2binding(
  annovar_path,
  vcf_path,
  genome_version = c("hg19", "hg38"),
  need_allsamples = FALSE,
  need_samples,
  mhc_type,
  pep_length,
  allele,
  pre_method
)
}
\arguments{
\item{annovar_path}{Character, the install path of annovar.}

\item{vcf_path}{Character, the path of VCF file needed to be converted.}

\item{genome_version}{Character, which genome build version of the VCF file, can be hg19 or hg38}

\item{need_allsamples}{Logical. Whether need all samples when multi-sample VCF file is supplied.}

\item{need_samples}{Character, the needed sample when multi-sample VCF file is supplied.}

\item{mhc_type}{MHC class, can be MHC-I or MHC-II.}

\item{pep_length}{A numeric or character vector, indicating the length for which to make predictions. For MHC-I, the length can be 8-15, for MHC-II, the length can be 11-30 or asis (take the length of input sequence as the peptide length)}

\item{allele}{A character vector of HLA alleles, available alleles for specific method can be obtained by \code{\link{available_alleles}}}

\item{pre_method}{Character, indicating the prediction method. Available methods for MHC-I or MHC-II can be obtained by \code{\link{available_methods}}}
}
\value{
A dataframe contains the predicted IC50 and precentile rank (if available).
}
\description{
This function use \code{\link{vcf2seq}} to exacted mutated "new peptide" from VCF file, and predicted the binding affinity of these peptide with specific MHC allele using \code{\link{general_mhcbinding}}
}
\examples{
test <- vcf2binding(annovar_path = "~/software/annovar/",vcf_path = system.file("extdata", "test_grch38.vcf", package = "MHCbinding"),
                    genome_version = "hg38",need_allsamples = FALSE,need_samples = "TUMOR",mhc_type = "MHC-I",pep_length = c(9,10),
                    allele = c("HLA-A*01:01", "HLA-A*03:01"),pre_method = "ann")
}
